fix(frontend): sqlserver数据迁移单据详情支持查看修改详情 #9802 # Reviewed, transaction id: 39890
