+++ dbm-ui/frontend/src/locales/zh-cn.json
@@ -3088,6 +3088,7 @@
   "构造后 DB 名": "构造后 DB 名",
   "手动修改回档的 DB 名": "手动修改回档的 DB 名",
   "n项待修改": "{0} 项待修改",
+  "n项已修改": "{0} 项已修改",
   "禁用的集群不支持提取 Key": "禁用的集群不支持提取 Key",
   "禁用的集群不支持删除 Key": "禁用的集群不支持删除 Key",
   "禁用的集群不支持备份": "禁用的集群不支持备份",
@@ -4358,5 +4359,8 @@
   "请先设置集群、构造 DB、回档信息": "请先设置集群、构造 DB、回档信息",
   "请先设置集群、目标集群、构造 DB、回档信息": "请先设置集群、目标集群、构造 DB、回档信息",
   "集群xx是已被选中的源集群": "集群{0}是已被选中的源集群",
+  "迁移 DB 名称": "迁移 DB 名称",
+  "迁移后 DB 名称": "迁移后 DB 名称",
+  "SQLServer数据迁移手动修改迁移DB名": "SQLServer数据迁移手动修改迁移DB名",
   "这行勿动！新增翻译请在上一行添加！": ""
 }
