+++ dbm-ui/frontend/src/views/ticket-center/common/ticket-detail/components/task-info/com-factory/sqlserver/common/data-migrate/components/rename-infos/components/ExportBtn.vue
@@ -13,33 +13,21 @@
 <script setup lang="ts">
   import { useI18n } from 'vue-i18n';
 
+  import TicketModel, { type Sqlserver } from '@services/model/ticket/ticket';
+
   import { exportExcelFile } from '@utils';
 
-  import type { IValue } from '../Index.vue';
-
   interface Props {
-    data: {
-      srcCluster: {
-        id: number;
-        master_domain: string;
-      };
-    };
+    data: TicketModel<Sqlserver.DataMigrate>['details']['infos'][number];
+    domain: string;
   }
 
   const props = defineProps<Props>();
 
-  const modelValue = defineModel<{
-    dbIgnoreName: string[];
-    dbName: string[];
-    renameInfoList: IValue[];
-  }>({
-    required: true,
-  });
-
   const { t } = useI18n();
   // 导出文件
   const handleExport = () => {
-    const formatData = modelValue.value.renameInfoList.map((item) => ({
+    const formatData = props.data.rename_infos.map((item) => ({
       [t('已存在的 DB')]: item.rename_db_name,
       [t('迁移 DB 名称')]: item.db_name,
       [t('迁移后 DB 名称')]: item.target_db_name,
@@ -49,8 +37,8 @@
     exportExcelFile(
       formatData,
       colsWidths,
-      `集群（${props.data.srcCluster.master_domain}）`,
-      `${t('SQLServer数据迁移手动修改迁移DB名')}_${props.data.srcCluster.master_domain}.xlsx`,
+      `集群（${props.domain}）`,
+      `${t('SQLServer数据迁移手动修改迁移DB名')}_${props.domain}.xlsx`,
     );
   };
 </script>
